add Spring Data JPA: connect PostgreSQL, refactor UserRepository

+++ src/main/java/com/github/niko91101/springpractice/controller/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable int id) {
+    public User getUser(@PathVariable long id) {
         return userService.getUserById(id);
     }
